Switch to use the strip ID as a document ID
There are a couple non-unique titles in the complete list of the
strips - meaning the assumption that the titles were kept unique is
plainly wrong.

So once backfilling missing strips was implemented, one would see the
same 4 strips overwriting each other at every run.

The most natural change to make is to use the strip ID as an index
instead. This adds a moderate amount of converting integers and strings
into each other.

Also, a special map of strips to skip was added to support cases of
skipped IDs - for now the only case is strip 404, which is
appropriately not found :P

Fixes #7